requirements.txt
asgiref==3.2.3
autopep8==1.4.4
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.13.2
cloudinary==1.19.1
cryptography==2.8
Django==3.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
fabric2==2.5.0
gunicorn==20.0.4
invoke==1.3.0
mock==3.0.5
paramiko==2.7.0
pkg-resources==0.0.0
psycopg2==2.8.4
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
PyNaCl==1.3.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
